Create src.html
The search begins around the start cell. We give this cell a value of 0. 
The next number to assign to a cell is 1
If the cell north of it is free we give it the value 1 and the next number to assign
to a cell is 2
We search the cells north, east, south and west of the current cell. Assigning
the next number to them
As every cell in the grid is search we reach stop cell. 
To find our path back to the start. We search again north, east, south
and west and find which of these cells has the lowest value.
This cell becomes part of the path and se search from here 
for the lowest value surrounding cell.